+++ src/game/Spells/SpellEntry.h
@@ -1069,16 +1069,7 @@ class SpellEntry
             return false;
         }
 
-        bool CanTriggerWeaponProcs() const
-        {
-            // All weapon based abilities can trigger weapon procs,
-            // even if they do no damage, or break on damage, like Sap.
-            // https://www.youtube.com/watch?v=klMsyF_Kz5o
-            if (EquippedItemClass == ITEM_CLASS_WEAPON && rangeIndex == SPELL_RANGE_IDX_COMBAT)
-                return true;
-
-            return HasAttribute(SPELL_CUSTOM_TRIGGER_WEAPON_PROCS);
-        }
+        bool CanTriggerWeaponProcs() const;
 
         bool HasDirectThreatIncreaseEffect() const
         {
